e2e fix

+++ Backend/src/ProzoroBanka.API/Controllers/PurchasesController.cs
@@ -161,7 +161,7 @@ public class PurchasesController : ApiControllerBase
 		[FromForm] IFormFile? file,
 		[FromForm] DocumentType type,
 		[FromForm] DateTime? documentDate,
-		[FromForm] long? amount,
+		[FromForm] decimal? amount,
 		[FromForm] string? counterpartyName,
 		CancellationToken ct)
 	{
